+++ open-codegen/CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Note: This is the Changelog file of `opengen` - the Python interface of OpEn
 
-## [Unreleased]
+## [0.4.1]
 
 ### Fixed
+
+* Project-specific `icasadi` dependency
+* Project-specific `tcp_iface` TCP interface
+* Fixed `lbfgs` typo
 
 ### Added
 
@@ -18,4 +22,4 @@ Note: This is the Changelog file of `opengen` - the Python interface of OpEn
 ### Removed
 
 
-[Unreleased]: https://github.com/alphaville/optimization-engine/compare/v0.6.1-alpha.2...master
+[0.4.1]: https://github.com/alphaville/optimization-engine/compare/opengen-0.4.1...master